get rid of compare-warnings

+++ src/LogicValue.cpp
@@ -549,9 +549,9 @@ bool LogicValue::operator>(LogicValue const& value)
     else if (_type == UIntType && value._type == UIntType)
         return (uintValue() > value.uintValue());
     else if (_type == UIntType)
-        return (uintValue() > value.intValue());
+        return ((int64_t)uintValue() > (int64_t)value.intValue());
     else if (value._type == UIntType)
-        return (intValue() > value.uintValue());
+        return ((int64_t)intValue() > (int64_t)value.uintValue());
     else
         return (intValue() > value.intValue());
 }
@@ -563,9 +563,9 @@ bool LogicValue::operator<(LogicValue const& value)
     else if (_type == UIntType && value._type == UIntType)
         return (uintValue() < value.uintValue());
     else if (_type == UIntType)
-        return (uintValue() < value.intValue());
+        return ((int64_t)uintValue() < (int64_t)value.intValue());
     else if (value._type == UIntType)
-        return (intValue() < value.uintValue());
+        return ((int64_t)intValue() < (int64_t)value.uintValue());
     else
         return (intValue() < value.intValue());
 }
@@ -577,9 +577,9 @@ bool LogicValue::operator>=(LogicValue const& value)
     else if (_type == UIntType && value._type == UIntType)
         return (uintValue() >= value.uintValue());
     else if (_type == UIntType)
-        return (uintValue() >= value.intValue());
+        return ((int64_t)uintValue() >= (int64_t)value.intValue());
     else if (value._type == UIntType)
-        return (intValue() >= value.uintValue());
+        return ((int64_t)intValue() >= (int64_t)value.uintValue());
     else
         return (intValue() >= value.intValue());
 }
@@ -591,9 +591,9 @@ bool LogicValue::operator<=(LogicValue const& value)
     else if (_type == UIntType && value._type == UIntType)
         return (uintValue() <= value.uintValue());
     else if (_type == UIntType)
-        return (uintValue() <= value.intValue());
+        return ((int64_t)uintValue() <= (int64_t)value.intValue());
     else if (value._type == UIntType)
-        return (intValue() <= value.uintValue());
+        return ((int64_t)intValue() <= (int64_t)value.uintValue());
     else
         return (intValue() <= value.intValue());
 }
